+++ src/PhpSpreadsheet/Calculation/Calculation.php
@@ -3729,6 +3729,8 @@ class Calculation
             //    Given two matrices of (potentially) unequal size, convert the larger in each dimension to match the smaller
             self::resizeMatricesShrink($operand1, $operand2, $matrix1Rows, $matrix1Columns, $matrix2Rows, $matrix2Columns);
         }
+        [$matrix1Rows, $matrix1Columns] = self::getMatrixDimensions($operand1);
+        [$matrix2Rows, $matrix2Columns] = self::getMatrixDimensions($operand2);
 
         return [$matrix1Rows, $matrix1Columns, $matrix2Rows, $matrix2Columns];
     }
@@ -4648,10 +4650,12 @@ class Calculation
             } elseif (!is_numeric($token) && !is_object($token) && isset(self::BINARY_OPERATORS[$token])) {
                 // if the token is a binary operator, pop the top two values off the stack, do the operation, and push the result back on the stack
                 //    We must have two operands, error if we don't
-                if (($operand2Data = $stack->pop()) === null) {
+                $operand2Data = $stack->pop();
+                if ($operand2Data === null) {
                     return $this->raiseFormulaError('Internal error - Operand value missing from stack');
                 }
-                if (($operand1Data = $stack->pop()) === null) {
+                $operand1Data = $stack->pop();
+                if ($operand1Data === null) {
                     return $this->raiseFormulaError('Internal error - Operand value missing from stack');
                 }