+++ src/PhpSpreadsheet/Calculation/TextData/Format.php
@@ -124,10 +124,12 @@ class Format
 
         $value = Helpers::extractString($value);
         $format = Helpers::extractString($format);
+        $format = (string) NumberFormat::convertSystemFormats($format);
 
         if (!is_numeric($value) && Date::isDateTimeFormatCode($format)) {
-            // @phpstan-ignore-next-line
-            $value = DateTimeExcel\DateValue::fromString($value) + DateTimeExcel\TimeValue::fromString($value);
+            $value1 = DateTimeExcel\DateValue::fromString($value);
+            $value2 = DateTimeExcel\TimeValue::fromString($value);
+            $value = (is_numeric($value1) && is_numeric($value2)) ? ($value1 + $value2) : (is_numeric($value1) ? $value2 : $value1);
         }
 
         return (string) NumberFormat::toFormattedString($value, $format);